Moved Save file to footer status bar.
