+++ src/main/resources/view/RootLayout.fxml
@@ -38,13 +38,7 @@
             </Menu>
          </menus>
       </MenuBar>
-      <AnchorPane styleClass="anchor-pane-with-border">
-         <children>
-            <Text fx:id="saveLocText" fill="WHITE" layoutX="14.0" layoutY="15.0" strokeType="OUTSIDE" strokeWidth="0.0" text="" AnchorPane.leftAnchor="10.0" />
-         </children>
-         <VBox.margin>
-            <Insets />
-         </VBox.margin>
+      <AnchorPane fx:id="footerStatusbarPlaceHolder" styleClass="anchor-pane-with-border">
       </AnchorPane>
       <SplitPane id="splitPane" fx:id="splitPane" prefHeight="295.0" prefWidth="209.0" VBox.vgrow="ALWAYS" />
    </children>
